backend/requirements.txt
# Web Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9

# Database
sqlalchemy>=2.0.25
aiosqlite>=0.19.0

# AI / LLM
openai>=1.12.0
tavily-python>=0.3.9

# Utilities
python-dotenv>=1.0.1
pydantic>=2.6.0
pydantic-settings>=2.1.0
httpx>=0.26.0
pandas>=2.2.0
